+++ src/LSST/M1M3/TS/Commands/SAL.h
@@ -62,6 +62,8 @@ SAL_COMMAND_CLASS_validate(MTM1M3TS, TSPublisher::SAL(), coolantPumpFrequency);
 
 SAL_COMMAND_CLASS(MTM1M3TS, TSPublisher::SAL(), coolantPumpReset);
 
+SAL_COMMAND_CLASS_validate(MTM1M3TS, TSPublisher::SAL(), applySetpoint);
+
 }  // namespace Commands
 }  // namespace TS
 }  // namespace M1M3