+++ src/LSST/M1M3/TS/Commands/Update.cpp
@@ -28,11 +28,17 @@
 
 #include "Commands/Update.h"
 
+#include "Events/AppliedSetpoint.h"
 #include "Events/EnabledILC.h"
+#include "Events/EngineeringMode.h"
 #include "Events/Heartbeat.h"
 #include "Events/SummaryState.h"
 #include "Events/ThermalInfo.h"
 
+#include <Settings/MixingValve.h>
+#include <Settings/Setpoint.h>
+
+#include "Telemetry/GlycolLoopTemperature.h"
 #include "Telemetry/MixingValve.h"
 #include "Telemetry/ThermalData.h"
 
@@ -49,6 +55,8 @@ LSST::cRIO::task_return_t Update::run() {
     _sendFCU();
 
     _sendMixingValve();
+
+    _temperatureControlLoop();
 
     Events::EnabledILC::instance().send();
 
@@ -96,7 +104,7 @@ void Update::_sendFCU() {
         Telemetry::ThermalData::instance().reset();
         TSApplication::ilc()->clear();
 
-        TSApplication::instance().callFunctionOnIlcs([](uint8_t address) -> void {
+        TSApplication::instance().callFunctionOnAllIlcs([](uint8_t address) -> void {
             if (Events::SummaryState::instance().active()) {
                 TSApplication::ilc()->reportThermalStatus(address);
             } else {
@@ -112,4 +120,51 @@ void Update::_sendFCU() {
     } catch (std::exception &e) {
         SPDLOG_WARN("Cannot poll FCU: {}", e.what());
     }
+}
+
+void Update::_temperatureControlLoop() {
+    auto timestep_ms = std::chrono::milliseconds(int(Settings::Setpoint::instance().timestep * 1000.0));
+
+    static auto next_update = std::chrono::steady_clock::now() - timestep_ms;
+
+    auto now = std::chrono::steady_clock::now();
+    if (now < next_update) {
+        return;
+    }
+
+    next_update += timestep_ms;
+
+    if (Events::SummaryState::instance().enabled() == false ||
+        Events::EngineeringMode::instance().isEnabled() == true) {
+        return;
+    }
+
+    auto mirrorLoop = Telemetry::GlycolLoopTemperature::instance().getMirrorLoopAverage();
+    float targetTemp = Events::AppliedSetpoint::instance().getAppliedSetpoint();
+    static float new_valve_position = 10.0;
+
+    float diff = mirrorLoop - targetTemp;
+    auto tolerance = Settings::Setpoint::instance().tolerance;
+
+    auto mixingValveStep = Settings::Setpoint::instance().mixingValveStep;
+
+    if (diff > tolerance) {
+        new_valve_position += mixingValveStep;
+    } else if (diff < -tolerance) {
+        new_valve_position -= mixingValveStep;
+    } else {
+        return;
+    }
+
+    if (new_valve_position > 100.0) {
+        new_valve_position = 100.0;
+    } else if (new_valve_position < 0) {
+        new_valve_position = 0;
+    }
+
+    SPDLOG_INFO("TemperatureControlLoop: new valve position is {:.1f}%, temperature difference was {:+.3f}",
+                new_valve_position, diff);
+
+    IFPGA::get().setMixingValvePosition(
+            Settings::MixingValve::instance().percentsToCommanded(new_valve_position));
 }